+++ playground/nuxt.config.ts
@@ -2,6 +2,9 @@ import { searchForWorkspaceRoot } from 'vite'
 
 export default defineNuxtConfig({
   modules: ['../src/module'],
+  // contextualTransition: {
+  //   hashScroll: false
+  // },
   vite: {
     server: {
       fs: {