+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog | Nuxt Contextual Transition 
 
+## [unreleased]
+
+### Added
+
+* You can now specify a `hashScroll` behavior in your `nuxt.config`  under `contextualTransition` to define what happens when you click on a link that leads you elsewhere on the same page. Options are `smooth` (default, smooth scroll), `instant`, `auto` to use the [css scroll behavior property](https://developer.mozilla.org/en-US/docs/Web/CSS/scroll-behavior), or `false` to turn it off.
+
+### Fixed
+
+* Compatibility set to Nuxt 3
+
 ## [0.1.0] - 2023-04-27 01:24PM
 
 ### Added